Share format writer helpers
